finishing main.cpp and removing unecessary files

+++ input.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include "Models.h"
 
-int getPosNum(const std::string& text);
+double getPosNum(const std::string& text);
 
 std::string getString(const std::string& text);
